pyproject.toml
[tool.poetry]
name = "open_precision"
version = "0.1.0"
description = ""
authors = ["larsjaeger <lars@jgr.one>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.22.4"
pyserial = "^3.5"
pyquaternion = "^0.9.9"
flatten-dict = "^0.4.2"
"ruamel.yaml" = "^0.17.21"
sparkfun-qwiic-icm20948 = "^0.0.1"
adafruit-python-shell = "^1.3.3"
adafruit-circuitpython-bno055 = "^5.4.1"
Flask = "^2.1.2"
Flask-SocketIO = "^5.2.0"
spidev = {version = "^3.5", optional = true}
gevent = "^21.12.0"
gevent-websocket = "^0.10.1"
dataclasses-json = "^0.5.7"
Flask-SQLAlchemy = "^2.5.1"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
